refactor: add a dedicated class option parser
